logback で Scala の Map object や case class を logging する時のやつを切り出し
