+++ src/main/scala/com/github/sudachi0114/logexample/LogExample.scala
@@ -13,25 +13,4 @@ object LogExample {
   logger.info("StructuredArguments.keyValue {}", keyValue("KEY", "VALUE"))
   logger.info("StructuredArguments.entries {}", entries(Map("k1" -> "v1", "k2" -> "v2").asJava))
   logger.info("StructuredArguments.array {}", array("array", "a", "b", "c"))
-
-  logger.info("map in map", value(
-    "Scala map",
-    Map(
-      "map" -> Map(
-        "list" -> List(1, 2, 3).asJava
-      ).asJava
-    ).asJava
-  ))
-
-  case class User(id: Long, name: String, interests: List[String])
-  val user = User(100, "userA", List("Scala", "Logging"))
-
-  logger.info("case class logging", value(
-    "case class user",
-    Map(
-      "id" -> user.id,
-      "name" -> user.name,
-      "interests" -> user.interests.asJava
-    ).asJava
-  ))
 }
