+++ src/test/scala/com/github/sudachi0114/logexample/CaseClassLoggingSpec.scala
@@ -1,0 +1,12 @@
+package com.github.sudachi0114.logexample
+
+import org.scalatest.funspec.AnyFunSpec
+
+class CaseClassLoggingSpec extends AnyFunSpec {
+  describe("logback case class test") {
+    it("should call CaseClassLogging") {
+      CaseClassLogging
+    }
+  }
+
+}
